test(state): move tests
